style: Renamed medel to model
