+++ backend/src/message/message.resolver.ts
@@ -1,5 +1,12 @@
-import { Args, ID, Parent, Query, ResolveField, Resolver } from '@nestjs/graphql';
-import { User } from 'src/users/models/user.medel';
+import {
+  Args,
+  ID,
+  Parent,
+  Query,
+  ResolveField,
+  Resolver,
+} from '@nestjs/graphql';
+import { User } from 'src/users/models/user.model';
 import { UsersService } from 'src/users/users.service';
 import { MessageService } from './message.service';
 import { DM, Message } from './model/message.model';
@@ -8,8 +15,8 @@ import { DM, Message } from './model/message.model';
 export class MessageResolver {
   constructor(
     private readonly messageService: MessageService,
-    private readonly usersService: UsersService
-    ) {}
+    private readonly usersService: UsersService,
+  ) {}
 
   /*
    ** ANCHOR: DM
@@ -18,7 +25,7 @@ export class MessageResolver {
   @Query((returns) => [DM], { nullable: true })
   async DM(
     @Args('user_id', { type: () => ID }) user_id: string,
-    @Args('other_user', { type: () => ID }) other_user: string
+    @Args('other_user', { type: () => ID }) other_user: string,
   ): Promise<DM[]> {
     return await this.messageService.getDM(user_id, other_user);
   }
@@ -27,15 +34,20 @@ export class MessageResolver {
    ** ANCHOR: ResolveField
    */
 
-   @ResolveField('messages', (returns) => [Message])
-   async messages(
-      @Parent() dm: DM,
-      @Args('offset') offset: number,
-      @Args('limit') limit: number
-    ): Promise<Message[]> {
-     const { user_id, other_id } = dm;
-     return await this.messageService.getMessages(user_id, other_id, offset, limit);
-   }
+  @ResolveField('messages', (returns) => [Message])
+  async messages(
+    @Parent() dm: DM,
+    @Args('offset') offset: number,
+    @Args('limit') limit: number,
+  ): Promise<Message[]> {
+    const { user_id, other_id } = dm;
+    return await this.messageService.getMessages(
+      user_id,
+      other_id,
+      offset,
+      limit,
+    );
+  }
 
   @ResolveField('other_user', (returns) => User)
   async getUser(@Parent() dm: DM): Promise<User> {
@@ -43,33 +55,19 @@ export class MessageResolver {
     return await this.usersService.getUserById(other_id);
   }
 
-   /*
+  /*
    ** ANCHOR: DMUSER
    */
 
-   @Query((returns) => [User])
-   async DmUsers(
+  @Query((returns) => [User])
+  async DmUsers(
     @Args('user_id', { type: () => ID }) user_id: string,
     @Args('offset') offset: number,
-    @Args('limit') limit: number
-     ): Promise<User[]> {
-       return await this.messageService.getDmUsers(user_id, offset, limit);
-   }
+    @Args('limit') limit: number,
+  ): Promise<User[]> {
+    return await this.messageService.getDmUsers(user_id, offset, limit);
+  }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 // @Resolver((of) => Message)
 // export class MessageResolver {
