+++ backend/src/message/model/message.model.ts
@@ -1,5 +1,5 @@
 import { Directive, Field, ID, Int, ObjectType } from '@nestjs/graphql';
-import { User } from 'src/users/models/user.medel';
+import { User } from 'src/users/models/user.model';
 
 @Directive('@key(fields: "user_id other_user { id }")')
 @ObjectType()
@@ -38,4 +38,3 @@ export class Message {
   @Field((type) => Int)
   date: number;
 }
-
